+++ src/agent_guard/__init__.py
@@ -10,6 +10,7 @@ from .context_guard import ContextGuardFinding, scan_context_files
 from .content_guard import ContentGuardFinding, scan_paths as scan_content_paths
 from .digest_guard import DigestGuardFinding, scan_digests
 from .path_guard import PathGuardFinding, scan_paths as scan_repo_paths
+from .workflow_guard import WorkflowGuardFinding, scan_workflow_policy
 
 scan_paths = scan_content_paths
 
@@ -25,6 +26,8 @@ __all__ = [
     "PathGuardFinding",
     "scan_digests",
     "DigestGuardFinding",
+    "scan_workflow_policy",
+    "WorkflowGuardFinding",
 ]
 
 __version__ = "0.1.3"